Switch out actions casting at Data layer

+++ curiosity/data/images_futures_and_actions.py
@@ -170,7 +170,7 @@ class FuturePredictionData(LMDBDataProvider): #LMDBDataProvider): # also uncomme
 	img, act, fut_img, fut_act, ids, fut_ids = self.create_image_pairs(batch[self.images], batch[self.actions])
 	
 	feed_dict = {'images': np.squeeze(img),
-		     'actions': np.squeeze(act).astype(np.float32),
+		     'actions': np.squeeze(act),
 		     'future_images': np.squeeze(fut_img),
 		     'future_actions': np.squeeze(fut_act)[:,0].astype(np.int32),
 		     'ids': np.squeeze(ids),
